refactor: extract mobile menu logic into separate HeaderMobileMenuWrapper component
