+++ src/components/Header/HeaderMobile/HeaderMobileMenuWrapper.tsx
@@ -1,0 +1,53 @@
+"use client";
+
+import { useEffect, useRef } from "react";
+import dynamic from "next/dynamic";
+import { usePathname } from "next/navigation";
+import { useDisclosure } from "@/hooks/useDisclosure";
+import HeaderMenuIcon from "@/icons/HeaderMenuIcon";
+import { cn } from "@/utils";
+
+const HeaderMobileMenu = dynamic(() =>
+  import("./HeaderMobileMenu").then((mod) => mod.HeaderMobileMenu),
+);
+
+export const HeaderMobileMenuWrapper = () => {
+  const { isOpen, toggle, close } = useDisclosure();
+  const menuButtonRef = useRef<HTMLButtonElement>(null);
+  const pathname = usePathname();
+
+  useEffect(() => {
+    close();
+  }, [pathname, close]);
+
+  const handleToggle = () => {
+    toggle();
+    if (!isOpen && menuButtonRef.current) {
+      menuButtonRef.current.focus();
+    }
+  };
+
+  return (
+    <>
+      <button
+        type="button"
+        data-testid="button"
+        aria-label="Otwórz menu"
+        onClick={handleToggle}
+        ref={menuButtonRef}
+      >
+        <HeaderMenuIcon className="self-center text-white/90 transition duration-300 hover:text-white active:scale-110 contrast:text-black00" />
+      </button>
+      <div
+        data-testid="mobile-menu"
+        className={cn(
+          "fixed right-0 top-0 z-50 h-full w-full transform transition-transform duration-300 ease-in-out",
+          "bg-greenB contrast:bg-yellowContrast",
+          isOpen ? "translate-x-0" : "translate-x-full",
+        )}
+      >
+        {isOpen && <HeaderMobileMenu onClose={close} />}
+      </div>
+    </>
+  );
+};
